gets the create action for blogpost working
updated the model to reflect what we want to keep in the database for
blogposts. uncommented out the timestamps and victuals for blogposts.
then writes the create script to get it working. also started to work on
the index but still like to test it more to make sure that its working
the way we want it too.